requirements.txt
# RSS parsing
feedparser>=6.0.10

# HTTP requests
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0

# Google Sheets
gspread>=5.12.0
google-auth>=2.23.0

# Bluesky API
atproto>=0.0.55

# Environment variables
python-dotenv>=1.0.0
